Nettoyage JourneyService et migration de la méthode searchJourney dans JourneySearchService.php
